Not working :/

+++ AskMessage.java
@@ -3,9 +3,9 @@ package com.exercise3;
 import java.util.Map;
 
 public class AskMessage {
-    final Map<Product, Integer> resources;
+    final Map<Product, Double> resources;
 
-    AskMessage(Map<Product, Integer> resources) {
+    AskMessage(Map<Product, Double> resources) {
         this.resources = resources;
     }
 }
