+++ TransferMessage.java
@@ -3,9 +3,9 @@ package com.exercise3;
 import java.util.Map;
 
 public class TransferMessage {
-    final Map<Product, Integer> resources;
+    final Map<Product, Double> resources;
 
-    TransferMessage(Map<Product, Integer> resources) {
+    TransferMessage(Map<Product, Double> resources) {
         this.resources = resources;
     }
 }
